Split handlers from bot
