+++ bin/bot.js
@@ -6,7 +6,49 @@
 */
 
 const config = require('../config');
-const CouchpotatoBot = require('../lib/bot');
+const MessageHandlers = require('../lib/handlers');
+const Bot = require('../lib/bot');
 
-let bot = new CouchpotatoBot(config.getProperties());
+let bot = new Bot(config.getProperties());
+
+bot.registerHandler({
+	group: 'channel',
+	match: /download movie\s(.*)/,
+	handler: MessageHandlers.downloadMovieHandler
+});
+
+bot.registerHandler({
+	group: 'channel',
+	match: /search movie\s(.*)/,
+	handler: MessageHandlers.searchMovieHandler
+});
+
+bot.registerHandler({
+	group: 'direct',
+	match: /hi/ig,
+	handler: function (message, callback) {
+		callback('Hi there! :smiley:');
+	}
+});
+
+bot.registerHandler({
+	group: 'direct',
+	match: /:popcorn:/ig,
+	handler: function (message, callback) {
+		callback('POPCORN! :yum:');
+	}
+});
+
+bot.registerHandler({
+	group: 'direct',
+	match: 'help',
+	handler: MessageHandlers.helpHandler
+});
+
+bot.registerHandler({
+	group: 'admin',
+	match: /!version/i,
+	handler: MessageHandlers.couchPotatoHandler
+});
+
 bot.run();
