+++ lib/bot.js
@@ -1,57 +1,29 @@
 'use strict';
-const Bot = require('slackbots');
-const _ = require('lodash');
-const MessageHandlers = require('./handlers');
 
-class CouchpotatoBot extends Bot {
+const Slackbots = require('slackbots');
+const _ = require('lodash');
+
+class Bot extends Slackbots {
 
 	constructor(settings) {
 		super(settings.bot);
-
 		this.settings = settings;
 		this.user = null;
-
-		this.cmds = {
-			channel: [
-				{
-					match: /download movie\s(.*)/,
-					handler: MessageHandlers.downloadMovieHandler
-				},
-				{
-					match: /search movie\s(.*)/,
-					handler: MessageHandlers.searchMovieHandler
-				}
-			],
-			direct: [
-				{
-					match: /hi/ig,
-					handler: function (message, callback) {
-						callback('Hi there! :smiley:');
-					}
-				},
-				{
-					match: /:popcorn:/ig,
-					handler: function (message, callback) {
-						callback('POPCORN! :yum:');
-					}
-				},
-				{
-					match: 'help',
-					handler: MessageHandlers.helpHandler
-				}
-			],
-			admin: [
-				{
-					match: /!version/i,
-					handler: MessageHandlers.couchPotatoHandler
-				}
-			]
-		};
+		this.cmds = {channel: [], direct: [], admin: []};
 	}
 
 	run() {
 		this.on('start', this._onStart);
 		this.on('message', this._onMessage);
+	}
+
+	registerHandler(handler) {
+		let group = handler.group;
+		if (group === 'channel' || group === 'direct' || group === 'admin') {
+			this.cmds[handler.group].push(handler);
+		} else {
+			throw new Error('Invalid handler group. Cannot register handler.');
+		}
 	}
 
 	_onStart() {
@@ -166,4 +138,4 @@ class CouchpotatoBot extends Bot {
 	}
 }
 
-module.exports = CouchpotatoBot;
+module.exports = Bot;
